starting to document the obsidian module
